breakdown the command file into individual files
